Move LCD and time functions into their own files
