+++ inc/Time.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include <cstdint>
+
+uint32_t Timer_100ns();
+uint32_t Timer_us();
+void Delay_100ns(uint32_t multiplier);
+void Delay_us(uint32_t delay);
